[M] Ongeki: Move model
