+++ src/main/java/icu/samnyan/aqua/sega/ongeki/OngekiRepos.kt
@@ -5,8 +5,8 @@ import icu.samnyan.aqua.net.games.GenericPlaylogRepo
 import icu.samnyan.aqua.net.games.GenericUserDataRepo
 import icu.samnyan.aqua.net.games.GenericUserMusicRepo
 import icu.samnyan.aqua.net.games.IUserRepo
+import icu.samnyan.aqua.sega.ongeki.model.*
 import icu.samnyan.aqua.sega.ongeki.model.gamedata.*
-import icu.samnyan.aqua.sega.ongeki.model.userdata.*
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
