+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetUserPreviewHandler.java
@@ -5,7 +5,7 @@ import icu.samnyan.aqua.sega.ongeki.OgkUserDataRepo;
 import icu.samnyan.aqua.sega.ongeki.OgkUserOptionRepo;
 import icu.samnyan.aqua.sega.general.BaseHandler;
 import icu.samnyan.aqua.sega.ongeki.model.response.GetUserPreviewResp;
-import icu.samnyan.aqua.sega.ongeki.model.userdata.UserData;
+import icu.samnyan.aqua.sega.ongeki.model.UserData;
 import icu.samnyan.aqua.sega.util.jackson.BasicMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
