+++ src/main/java/icu/samnyan/aqua/sega/ongeki/handler/impl/GetUserRecentRatingHandler.java
@@ -5,8 +5,8 @@ import icu.samnyan.aqua.sega.general.model.response.UserRecentRating;
 import icu.samnyan.aqua.sega.ongeki.OgkUserGeneralDataRepo;
 import icu.samnyan.aqua.sega.ongeki.OgkUserPlaylogRepo;
 import icu.samnyan.aqua.sega.general.BaseHandler;
-import icu.samnyan.aqua.sega.ongeki.model.userdata.UserGeneralData;
-import icu.samnyan.aqua.sega.ongeki.model.userdata.UserPlaylog;
+import icu.samnyan.aqua.sega.ongeki.model.UserGeneralData;
+import icu.samnyan.aqua.sega.ongeki.model.UserPlaylog;
 import icu.samnyan.aqua.sega.util.jackson.BasicMapper;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
